refactor(theme): extract utility functions and improve modularity in color handling
- Moved `DynamicColorKey`, `getCurve`, `tMaxC`, `tMinC`, and related logic from `default-color` to new `color.utils` for improved modularity and reusability.
- Updated references across `color.manager`, `color.api`, and other modules to use `color.utils` imports.
- Refined `themer.ts` by replacing direct `themes` and `colors` usage with `api` for better abstraction and readability.
